clean code, delete mail file, rename file
